+++ arrow-cast/Cargo.toml
@@ -48,9 +48,5 @@ num = { version = "0.4", default-features = false, features = ["std"] }
 lexical-core = { version = "^0.8", default-features = false, features = ["write-integers", "write-floats", "parse-integers", "parse-floats"] }
 
 [dev-dependencies]
-half = { version = "2.1", default-features = false, features = ["num-traits"] }
-arrow-array = { version = "28.0.0", path = "../arrow-array", features = ["chrono-tz"]}
-
-[features]
 
 [build-dependencies]
